requirements.txt
aiohttp==3.9.3
azure-identity==1.18.0
python-dotenv==1.0.1